image-pipeline/requirements.txt
python-dotenv
pillow
requests
pyyaml
